Fixed AutoMapper

+++ Garage2/Program.cs
@@ -30,8 +30,11 @@ public class Program
 	        var serviceProvider = scope.ServiceProvider;
 	        var db = serviceProvider.GetRequiredService<Garage2Context>();
 
-	        //erases the database everytime i run the application
-	        await db.Database.EnsureDeletedAsync();
+	        //If there isn't any parkedVehicles in the database erase it and instantiate defaults
+	        if (!db.ParkedVehicle.Any())
+	        {
+		        await db.Database.EnsureDeletedAsync();
+			}
 	        //run all the migrations, if the database doesnt exist create it, if it exist, just update the database
 	        await db.Database.MigrateAsync();
